Fixed incompatibility with WP 3.3+ due to a change in the way a WP meta function works
git-svn-id: https://plugins.svn.wordpress.org/contexture-page-security/trunk@506147 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ core/CTXPS_Ajax.php
@@ -380,7 +380,7 @@ class CTXPS_Ajax {
                 $response = array(
                     'what'=>    'update_sec',
                     'action'=>  'update_security',
-                    'id'=>      add_metadata($_REQUEST['object_type'], $_REQUEST['object_id'], 'ctx_ps_security', '1', true),
+                    'id'=>      (int)add_metadata($_REQUEST['object_type'], $_REQUEST['object_id'], 'ctx_ps_security', '1', true),
                     'data'=>    __('Security enabled.','contexture-page-security')
                 );
                 break;
